rename locking.js to wallet.js sagas
